+++ R/Interplot_mlm.R
@@ -227,6 +227,19 @@ interplot.lmerMod <- function(m, var1, var2, plot = TRUE, steps = NULL, ci = .95
                   unlist(dimnames(m@pp$X)[2]))], 1 - (1 - ci) / 2)
         }
         
+        multiplier <- if (var1 == var2) 
+          2 else 1
+        
+        min_sim <- m.sims@fixef[, match(var1, unlist(dimnames(m@pp$X)[2]))] + 
+          multiplier * xmin * m.sims@fixef[, match(var12, unlist(dimnames(m@pp$X)[2]))] # simulation of the value at the minimum value of the conditioning variable
+        max_sim <- m.sims@fixef[, match(var1, unlist(dimnames(m@pp$X)[2]))] + 
+          multiplier * xmax * m.sims@fixef[, match(var12, unlist(dimnames(m@pp$X)[2]))] # simulation of the value at the maximum value of the conditioning variable
+        diff <- max_sim - min_sim # calculating the difference
+        ci_diff <- c(
+          quantile(diff, (1 - ci) / 2),
+          quantile(diff, 1 - (1 - ci) / 2)
+        ) # confidence intervals of the difference
+        
         if (plot == TRUE) {
             if (hist == TRUE) {
                 if (is.na(var2_dt)) {
@@ -236,8 +249,7 @@ interplot.lmerMod <- function(m, var1, var2, plot = TRUE, steps = NULL, ci = .95
                 }
             }
             interplot.plot(m = coef, hist = hist, var2_dt = var2_dt, point = point, 
-                ercolor = ercolor, esize = esize, ralpha = ralpha, rfill = rfill, 
-                ...)
+                ercolor = ercolor, esize = esize, ralpha = ralpha, rfill = rfill, ci_diff = ci_diff, ...)
         } else {
             names(coef) <- c(var2, "coef", "ub", "lb")
             return(coef)
@@ -521,6 +533,19 @@ interplot.glmerMod <- function(m, var1, var2, plot = TRUE, steps = NULL, ci = .9
           }
         }
         
+      multiplier <- if (var1 == var2) 
+        2 else 1
+      
+      min_sim <- m.sims@fixef[, match(var1, unlist(dimnames(m@pp$X)[2]))] + 
+        multiplier * xmin * m.sims@fixef[, match(var12, unlist(dimnames(m@pp$X)[2]))] # simulation of the value at the minimum value of the conditioning variable
+      max_sim <- m.sims@fixef[, match(var1, unlist(dimnames(m@pp$X)[2]))] + 
+        multiplier * xmax * m.sims@fixef[, match(var12, unlist(dimnames(m@pp$X)[2]))] # simulation of the value at the maximum value of the conditioning variable
+      diff <- max_sim - min_sim # calculating the difference
+      ci_diff <- c(
+        quantile(diff, (1 - ci) / 2),
+        quantile(diff, 1 - (1 - ci) / 2)
+      ) # confidence intervals of the difference
+      
         if (plot == TRUE) {
             if (hist == TRUE) {
                 if (is.na(var2_dt)) {
@@ -529,7 +554,7 @@ interplot.glmerMod <- function(m, var1, var2, plot = TRUE, steps = NULL, ci = .9
                   var2_dt <- var2_dt
                 }
             }
-            interplot.plot(m = coef, steps = steps, hist = hist, predPro = predPro, var2_vals = var2_vals, var2_dt = var2_dt, point = point, ercolor = ercolor, esize = esize, ralpha = ralpha, rfill = rfill, ...)
+            interplot.plot(m = coef, steps = steps, hist = hist, predPro = predPro, var2_vals = var2_vals, var2_dt = var2_dt, point = point, ercolor = ercolor, esize = esize, ralpha = ralpha, rfill = rfill, ci_diff = ci_diff, ...)
         } else {
             if(predPro == TRUE){
               names(coef) <- c(var2, paste0("values_in_", var1), "coef", "ub", "lb")